backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.9.0
python-dotenv==1.0.0
pydantic==2.5.0
anthropic==0.40.0
# Auth dependencies
python-jose[cryptography]==3.3.0
httpx==0.27.0